pyproject.toml
[project]
name = "sentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apache-airflow>=3.1.8",
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "goose3>=3.1.21",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "newspaper4k>=0.9.4.1",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "streamlit>=1.55.0",
    "textblob>=0.19.0",
    "trafilatura>=2.0.0",
    "vadersentiment>=3.3.2",
    "wordcloud>=1.9.6",
]
